Clear the current interval when no more intervals can be resolved
When no more intervals can be resolved, ResolveInterval(...) will return
an interval that return false from IsResolved() (meaning it has no
finite begin time). When this happens we should "clear" |interval_| by
setting it to an unresolved interval.

This fixes the behavior of SVGAnimationElement.getStartTime() which now
correctly report (throws) when the last interval of element has ended.
Add a new test for that and move an old somewhat related test to WPT.
Also adjust SVGAnimationTestCase.js which uses this to offset the sample
times (although in most cases) to avoid the debug dump when there's no
current interval (no tests actually rely on the old broken behavior).

This makes us match Firefox.

Bug: 998526
Change-Id: Idcf348343768ffc9392e29af2e976b9a3a1e31aa